LayoutTests/fast/css/parsing-stroke-linejoin-expected.txt

This tests checks that all of the input values for stroke-linejoin parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyleInherited("stroke-linejoin: miter;") is "miter"
PASS testComputedStyleInherited("stroke-linejoin: round;") is "round"
PASS testComputedStyleInherited("stroke-linejoin: bevel;") is "bevel"
PASS testComputedStyle(";") is "miter"
PASS test("stroke-linejoin: miter;") is "miter"
PASS test("stroke-linejoin: round;") is "round"
PASS test("stroke-linejoin: bevel;") is "bevel"
PASS test("stroke-linejoin: mitr;") is ""
PASS test("stroke-linejoin: bevl;") is ""
PASS test("stroke-linejoin: 10px;") is ""
PASS test("stroke-linejoin: 10%;") is ""
PASS successfullyParsed is true

TEST COMPLETE
